+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.7.0
+
+  * Breaking Change: Actions you emit from your Epic are now re-dispatched through all Epics. They used to be simply forwarded to the next Middleware in the chain.  
+  * Added support for `async*` functions
+  * Added `TypedEpic` as a convenient way to bind actions of a certain type to an epics.
+
 # 0.6.1
 
 Improve docs, bump to ensure it works with latest RxDart.